% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyAnalysis.r
\name{addAnnualXlWorksheet}
\alias{addAnnualXlWorksheet}
\title{Add Annual Excel Worksheet}
\usage{
addAnnualXlWorksheet(data_list, data_name, result_doc, worksheet_name)
}
\arguments{
\item{data_list}{with monthly survey resultss}

\item{data_name}{name of data frames to combine}

\item{result_doc}{An excel document}

\item{worksheet_name}{The name of the work sheet to add}
}
\value{
The list provided as data_list is returned (allows writing mulitple worksheets in a pipe)
}
\description{
Helper funciton to combine list of monthly data frames into a single annual data frames
and writes it to a provided Excel document as a new worksheet
}
