% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyAnalysisLib.r
\name{addDrawType}
\alias{addDrawType}
\title{Add Draw Type to survey results (e.g. Pre or Post)}
\usage{
addDrawType(licence_data, survey_dates)
}
\arguments{
\item{licence_data}{Data frame of licences}

\item{survey_dates}{Electronic licences data, including specifics of each licence}
}
\value{
The licence data frame with an additional draw_type column
}
\description{
Adds the draw type based on selection criteria and survey dates for
each licence record
}
