% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyAnalysisLib.r
\name{addPurchasePeriod}
\alias{addPurchasePeriod}
\title{Add Purchase Periods to licence records (e.g. In-Survey, Pre-Survey, Early-Season, and Late-Season)}
\usage{
addPurchasePeriod(
  licence_data,
  survey_dates,
  period_stratify,
  period_stratify_date = NA
)
}
\arguments{
\item{licence_data}{Data frame of licences}

\item{survey_dates}{Electronic licences data, including specifics of each licence}

\item{period_stratify}{Boolean identifying if period stratification is enabled}

\item{period_stratify_date}{If splitting into early/late season, this identifies date of split}
}
\value{
The licence data frame with an additional draw_type column
}
\description{
Adds the purchase period identifiers to each licence record based on the if period stratification is
enabled (TRUE) and identify early/late season licences based on the period stratification date
}
