% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyAnalysisLib.r
\name{addXlWorksheet}
\alias{addXlWorksheet}
\title{Add Excel Worksheet}
\usage{
addXlWorksheet(data_table, excel_doc, sheet_name)
}
\arguments{
\item{data_table}{Data that is to be written to the new worksheet}

\item{excel_doc}{The excel document variable created with openxlsx::createworkbook()}

\item{sheet_name}{Name that the new worksheet is to be labelled as}
}
\value{
The data table provide, unmodified
}
\description{
This AddWorksheet function adds a worksheet to a previously create excel document
}
\details{
The data table is the first parameter and it return sthe first parameter unchanged
so that it is compatible with magrittr "\%>\%"
}
