% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyAnalysisLib.r
\name{calcResponseTotals}
\alias{calcResponseTotals}
\title{Calculate Response Totals}
\usage{
calcResponseTotals(
  survey_data,
  uniq_user_col_name = "licence_id",
  strata_col_names = c("draw_type", "resident_status", "licence_type", "age_category")
)
}
\arguments{
\item{survey_data}{The survey response data}

\item{uniq_user_col_name}{Column name for unique users}

\item{strata_col_names}{Stratification column names}
}
\value{
the survey_data with additional columns that match the catch columns with the "n_" prefix
}
\description{
Add up responce total by strata and catch category.
Because adjusments may remove catch from a particular user, then each catch
cateogry may have a seperate response total.
}
