% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyUtil.r
\name{convertColTypes}
\alias{convertColTypes}
\title{Convert Column Types}
\usage{
convertColTypes(df, df_to_match)
}
\arguments{
\item{df}{The tibble whose columns are to be converted}

\item{df_to_match}{The tibble with the column type definitions}
}
\value{
The df tibble with the column types converted to match the df_to_match tibble
}
\description{
Convert column types of a data frame to match another tibble with the same column names.
This function helps to setup tibbles to use dplyr::bind_rows with
}
\details{
NOTE: the column names must be exactly the same between the two tibbles
}
