% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSurveyAdjFile.r
\name{createAdjFile}
\alias{createAdjFile}
\title{Create Adjustment File}
\usage{
createAdjFile(adj_lic_id, config)
}
\arguments{
\item{adj_lic_id}{licences ids to export in the adjustment file}

\item{config}{Configuration list}
}
\value{
The file name that the adjustment file is created
}
\description{
Create an adjustment file based on specific licence ids
}
