% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportInstRaw.r
\name{exportInstRawResult}
\alias{exportInstRawResult}
\title{Export Instantaniace Survey Raw Results}
\usage{
exportInstRawResult(
  licence_filename,
  survey_result_filename,
  output_data_filename
)
}
\arguments{
\item{licence_filename}{Licence file name}

\item{survey_result_filename}{Survey result file name}

\item{output_data_filename}{Output file name}
}
\value{
The Excel document file name that contains all the annual data
}
\description{
Export the raw survey results for an instantaniace survey
}
