% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyAnalysis.r
\name{exportSingleMonthResults}
\alias{exportSingleMonthResults}
\title{Export Single Month Survey Results}
\usage{
exportSingleMonthResults(
  lic_year,
  month_name,
  irec_dir_root = getiRecAnalysisDir()
)
}
\arguments{
\item{lic_year}{Licence year (e.g. 2016 for licence year 2016/17)}

\item{month_name}{Month to analyze taken from month.name (e.g "February")}

\item{irec_dir_root}{Path to the iRec data root directory}
}
\value{
The Excel document file name that contains all the month's survey responses
}
\description{
Combine the survey results with licence category information and export to an excel file.
This includes applying all adjustments.
}
