% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyUtil.r
\name{getiRecAnalysisDir}
\alias{getiRecAnalysisDir}
\title{Get the iRecAnalysis directory}
\usage{
getiRecAnalysisDir()
}
\value{
Root directory of iRecAnalysis data
}
\description{
Get the iRecAnalysis directory
}
