% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyDb.r
\name{getLicenceYearEnd}
\alias{getLicenceYearEnd}
\title{Get Licence Year End}
\usage{
getLicenceYearEnd(survey_start_date)
}
\arguments{
\item{survey_start_date}{Date a survey is carried out for.}
}
\value{
A end date for the licence year that the survey_start_date falls within
}
\description{
Helper method to get the ending date for the a licence year, inferred from provided date.
}
