% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyDb.r
\name{getLicenceYearStart}
\alias{getLicenceYearStart}
\title{Get Licence Year Start}
\usage{
getLicenceYearStart(survey_start_date)
}
\arguments{
\item{survey_start_date}{Date a survey is carried out for.}
}
\value{
A start date for the licence year that the survey_start_date falls within
}
\description{
Helper method to get the starting date for the a licence year, inferred from provided date.
}
