% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyDb.r
\name{getLicYearStartDate}
\alias{getLicYearStartDate}
\title{Get Licence Year Start Date}
\usage{
getLicYearStartDate(lic_end_date)
}
\arguments{
\item{lic_end_date}{Date a licence year ends}
}
\value{
Start date for the licence year
}
\description{
Helper method to get the starting date for the a licence year,
inferred from provided licence year end date.
}
