% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyUtil.r
\name{getLogMessages}
\alias{getLogMessages}
\title{Get Log Messages}
\usage{
getLogMessages()
}
\value{
A tibble with the log messages as a column
}
\description{
Retrieve the global vector of vector log messages
}
