% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyUtil.r
\name{getMonthEndDate}
\alias{getMonthEndDate}
\title{Get Month End Date}
\usage{
getMonthEndDate(date_param)
}
\arguments{
\item{date_param}{Date with month that you are requesting the end date for.}
}
\value{
A date for the last day of the month
}
\description{
A helper function for getting the end date for the month of the date provided
}
