% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NrlsDb.r
\name{getNrlsLicences}
\alias{getNrlsLicences}
\title{Get NRLS Licences}
\usage{
getNrlsLicences(db_conn, lic_year)
}
\arguments{
\item{db_conn}{Database connection to NRLS}

\item{lic_year}{Licence year to retrieve}
}
\value{
A data frame of licence records
}
\description{
Gets licence records from NRLS
}
