% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyInstAnalysis.r
\name{getNrlsPopSize}
\alias{getNrlsPopSize}
\title{Get NRLS Population Sizes}
\usage{
getNrlsPopSize(elic_data, lic_id_col_name, lic_strata_col_names)
}
\arguments{
\item{elic_data}{Electronic licence data frame}

\item{lic_id_col_name}{Name of column that identifies the licence ID}

\item{lic_strata_col_names}{Column names that are used to stratify for population totals}
}
\value{
A data frame population sizes based on strata columns
}
\description{
Summarize the E-Licence data to population sizes by licence stratification.
}
