% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyAnalysisLib.r
\name{getSampSizeColNames}
\alias{getSampSizeColNames}
\title{Sample Size Column Names}
\usage{
getSampSizeColNames(survey_df)
}
\arguments{
\item{survey_df}{A survey related data frame}
}
\value{
A vector of column names related survey sample sizes, including effort_days and juv_effort_days
}
\description{
Retrieve a vector of sample size column names that matches the catch column name vector length
}
