% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyUtil.r
\name{getSurveyPath}
\alias{getSurveyPath}
\title{Get Survey Path}
\usage{
getSurveyPath(month_name, year_data_path, annual_expire_date)
}
\arguments{
\item{month_name}{Name of month for the survey}

\item{year_data_path}{File path to the year data}

\item{annual_expire_date}{Date that the annual licence expires (e.g. March 31st)}
}
\value{
An absolute file path where the survey data should be available
}
\description{
Get the survey path based on survey parameters.
}
