% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyUtil.r
\name{getSurveyYearPath}
\alias{getSurveyYearPath}
\title{Get Survey Year Path}
\usage{
getSurveyYearPath(lic_year, irec_dir_root = getiRecAnalysisDir())
}
\arguments{
\item{lic_year}{Licence year text (e.g. "2013-14")}

\item{irec_dir_root}{Root directory for iRec Analysis}
}
\value{
An absolute file path where the survey year data should be available
}
\description{
Get the survey year path based on survey parameters.
}
