% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyUtil.r
\name{GetTimeStampText}
\alias{GetTimeStampText}
\title{Get Time Stamp Text}
\usage{
GetTimeStampText()
}
\value{
A string with the current date and time
}
\description{
(Deprecated) Provides a standardized text based time stamp for inclusion in file names
}
