% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyUtil.r
\name{getTruncText}
\alias{getTruncText}
\title{Get Truncated Text}
\usage{
getTruncText(values, show_n = 10, collapse = "\\n")
}
\arguments{
\item{values}{A vector of values to turn into text}

\item{show_n}{How many elements to show before truncating}

\item{collapse}{String used to combine the multiple values}
}
\value{
Text with the a vector of values to print to a log file
}
\description{
Provides a formatted chunk of text to display to the user by limiting output to the first
'show_n' elements for the vector of 'values' and providing text identifying how many more
are in the vector
}
\examples{
getTruncText(letters)

}
