% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyUtil.r
\name{labelText}
\alias{labelText}
\title{Convert SPSS labels to text}
\usage{
labelText(values)
}
\arguments{
\item{values}{A vector of labelled values}
}
\value{
Text vector
}
\description{
Columns read from an SPSS file with the haven R package provides labelled columns.
This function converts the labels (that are kind of like factors) into a text column.
}
