% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyUtil.r
\name{loadAnalysisConfig}
\alias{loadAnalysisConfig}
\title{Load Analysis Configuration}
\usage{
loadAnalysisConfig(
  lic_year,
  month_name,
  year_config_filename = "year_config.yml",
  survey_config_filename = "survey_config.yml",
  irec_dir_root = getiRecAnalysisDir(),
  year_config = NULL
)
}
\arguments{
\item{lic_year}{Licence year text to load configuration for (e.g "2016-17")}

\item{month_name}{Survey month name to load survey configuration for (e.g. "August")}

\item{year_config_filename}{Path to the year configuration file relative to survey config}

\item{survey_config_filename}{Survey configuration file name}

\item{irec_dir_root}{Root directory will iRec data}

\item{year_config}{Year configuration list if it has already been loaded

NOTE: For iArc surveys, the month_name is "iArc"}
}
\value{
A list of configuration parameters
}
\description{
Load the Year and Survey specific configuration file, adjust file names to full paths
}
