% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyUtil.r
\name{loadAnalysisYearConfig}
\alias{loadAnalysisYearConfig}
\title{Load Analysis Year Config}
\usage{
loadAnalysisYearConfig(
  lic_year,
  year_config_filename = "year_config.yml",
  irec_dir_root = getiRecAnalysisDir()
)
}
\arguments{
\item{lic_year}{Licence year text to load configuration for (e.g "2016-17")}

\item{year_config_filename}{Path to the year configuration file relative to survey config}

\item{irec_dir_root}{The root directory of the iRec Survey Software}
}
\value{
A list of configuration parameters
}
\description{
Load the Year configuration file, adjust file names to full paths
}
