% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyDb.r
\name{loadEkosSurveyResults}
\alias{loadEkosSurveyResults}
\title{Load EKOS Survey Results}
\usage{
loadEkosSurveyResults(ekos_filename, exclude_lic_id)
}
\arguments{
\item{ekos_filename}{File name of the Ekos survey results}

\item{exclude_lic_id}{A vector licences IDs to exclude from the results}
}
\value{
The survey results loaded from an Ekos file.
}
\description{
Load survey results from a EKOS survey file format
}
