% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyDb.r
\name{loadExcludeFile}
\alias{loadExcludeFile}
\title{Load Exclude File}
\usage{
loadExcludeFile(exclude_file_name = NA)
}
\arguments{
\item{exclude_file_name}{File name of excluded licences or survey access code}
}
\value{
A data frame of licence_id or survey_access_key to exclude
}
\description{
Read in the exclude csv file of licence IDs or Survey Keys that are to be
excluded from iRec Survey
}
