% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyDb.r
\name{loadLicenceFile}
\alias{loadLicenceFile}
\title{Load Licence File}
\usage{
loadLicenceFile(licence_filename, lic_start_date, lic_end_date)
}
\arguments{
\item{licence_filename}{File name that licence records are saved in CSV format}

\item{lic_start_date}{Licence year start date}

\item{lic_end_date}{Licence year end date}
}
\value{
A data frame with the licence records
}
\description{
Based on the licence start year, this function reads the appropriate licence file
format
}
