% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NrlsDb.r
\name{loadNrlsLicenceFile}
\alias{loadNrlsLicenceFile}
\title{Load NRLS Licence File}
\usage{
loadNrlsLicenceFile(licence_filename, lic_start_date, lic_end_date)
}
\arguments{
\item{licence_filename}{File name that licence records are saved in CSV format}

\item{lic_start_date}{Licence start date}

\item{lic_end_date}{Licence end date}
}
\value{
A data frame with the licence records
}
\description{
The main method used to load the licence data from NRLS system.  This function selects the
appropriate version of the file format to load based on the survey year
(starting April 1st)
}
