% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyDb.r
\name{loadPraSurveyResults}
\alias{loadPraSurveyResults}
\title{Loads the survey results from PRA and merging the data into common format
for analysis}
\usage{
loadPraSurveyResults(survey_result_filename, exclude_id, strata_col_names)
}
\arguments{
\item{survey_result_filename}{The file name that the PRA survey results saved too}

\item{exclude_id}{A vector licence_id(s) or survey_access_key(s) that should have their survey results
excluded from analysis}

\item{strata_col_names}{Strata and analysis columns to preserve in loaded file}
}
\value{
A tibble with the survey results
}
\description{
Loads the survey results from PRA and merging the data into common format
for analysis
}
