% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyDb.r
\name{loadVendorSalesData}
\alias{loadVendorSalesData}
\title{Load Vendor Sales}
\usage{
loadVendorSalesData(vendor_sales_filename, survey_month, annual_expire_date)
}
\arguments{
\item{vendor_sales_filename}{The file name that has vendors sales by month for all 12 months of the licence year.}

\item{survey_month}{The month name of the current month that is being analyzed}

\item{annual_expire_date}{The end date of the current annual licence year.}
}
\value{
A data frame with the vendor sales summarized to the licence types for before- and in-month sales.
}
\description{
Loads the full set of vendor sales totals by month and licence categories (e.g. annual/term).
Based on the survey month and the annual licence expire date, the months total sales are summarized into
those that occured before the month and in the month.
}
