% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyAnalysisLib.r
\name{mergeELicSurveyData}
\alias{mergeELicSurveyData}
\title{Merge Electronic Licence Survey Data}
\usage{
mergeELicSurveyData(survey_data, elic_data)
}
\arguments{
\item{survey_data}{Survey results data frame}

\item{elic_data}{Electronic licences data, including specifics of each licence}
}
\value{
The survey results with the licence details attached to each survey result record
}
\description{
Merges the survey data with licence data so that the survey results
will have the licence category information.  The two data sets are merged using
"licence_id"
}
