% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyUtil.r
\name{parseDate}
\alias{parseDate}
\title{Parse dates.  This function tries to parse using the ymd then myd and finally dmy format.}
\usage{
parseDate(date_text)
}
\arguments{
\item{date_text}{Dates to parse}
}
\description{
Parse dates.  This function tries to parse using the ymd then myd and finally dmy format.
}
