% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSurveyAdjFile.r
\name{parseLicenceIds}
\alias{parseLicenceIds}
\title{Parse user provided licence ID text based on
comma or space delimited list of Licence IDs}
\usage{
parseLicenceIds(licence_id_txt)
}
\arguments{
\item{licence_id_txt}{Text of licence IDs to parse}
}
\value{
character vector of unique licences IDs or an empty vector if on were provided
}
\description{
Parse user provided licence ID text based on
comma or space delimited list of Licence IDs
}
