% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NrlsDb.r
\name{refreshNrlsLicenceFile}
\alias{refreshNrlsLicenceFile}
\title{Refresh NRLS licence file}
\usage{
refreshNrlsLicenceFile(
  lic_year,
  config = NULL,
  irec_dir_root = getiRecAnalysisDir()
)
}
\arguments{
\item{lic_year}{Licence year to refresh licence records for}

\item{config}{Licence year configuration list}

\item{irec_dir_root}{Root directory that contains iRec Analysis}
}
\value{
File name that was refreshed
}
\description{
Connect with NRLS database and retrieve the latest licence records
}
