% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyAnalysis.r
\name{runAnnualAnalysis}
\alias{runAnnualAnalysis}
\title{Run Annual Analysis}
\usage{
runAnnualAnalysis(lic_year, irec_dir_root = getiRecAnalysisDir())
}
\arguments{
\item{lic_year}{Licence year (e.g. 2016 for licence year 2016/17)}

\item{irec_dir_root}{Path to the iRec data root directory}
}
\value{
The Excel document file name that contains all the annual data
}
\description{
Run analysis for all months within a licence year to a single annual Excel file
}
