% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyInstAnalysis.r
\name{runInstSurveyAnalysis}
\alias{runInstSurveyAnalysis}
\title{Run Instantaniace selection survey analysis}
\usage{
runInstSurveyAnalysis(config, elic_data, stamp_stratify = FALSE)
}
\arguments{
\item{config}{Configuration list}

\item{elic_data}{Electronic licence data frame}

\item{stamp_stratify}{Stratify based on salmon stamp}
}
\value{
list with survey analysis results
}
\description{
Analyse the responses to the iRec survey based on instantaniace selection method
integrated into the National Recreational Licence System (NRLS)
}
