% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyPrePostAnalysis.r
\name{runPrePostSurveyAnalysis}
\alias{runPrePostSurveyAnalysis}
\title{Run Pre/Post Survey Analysis}
\usage{
runPrePostSurveyAnalysis(config, elic_data)
}
\arguments{
\item{config}{Survey configuration list}

\item{elic_data}{Electronic lience data}
}
\value{
The list provided as data_list is returned (allows writing multiple worksheets in a pipe)
}
\description{
Processes survey results based on the pre/post sampled licences
(licence years 2012 to 2017)
}
