% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyAnalysis.r
\name{runSingleMonthAnalysis}
\alias{runSingleMonthAnalysis}
\title{Run Single Month Analysis}
\usage{
runSingleMonthAnalysis(
  lic_year,
  month_name,
  irec_dir_root = getiRecAnalysisDir()
)
}
\arguments{
\item{lic_year}{Licence year (e.g. 2016 for licence year 2016/17)}

\item{month_name}{Month to analyse take from month.name (e.g "February")}

\item{irec_dir_root}{Path to the iRec data root directory}
}
\value{
The Excel document file name that contains all the month's survey data
}
\description{
Run analaysis for single survey month and write results to an Excel file
}
