% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyAnalysis.r
\name{runUI}
\alias{runUI}
\title{Run iRecAnalysis User Interface}
\usage{
runUI(analysis_dir, port = 9001)
}
\arguments{
\item{analysis_dir}{Directory where iRec analysis is maintained}

\item{port}{Port number to run shiny server on}
}
\description{
Start a shiny server with the user interface for the iRecAnalysis package
}
