% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyUtil.r
\name{setiRecAnalysisDir}
\alias{setiRecAnalysisDir}
\title{Set iRecAnalysis Directory}
\usage{
setiRecAnalysisDir(dir)
}
\arguments{
\item{dir}{Root directory of iRecAnalysis data}
}
\description{
Set iRecAnalysis Directory
}
