% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NrlsDb.r
\name{setupNrlsConn}
\alias{setupNrlsConn}
\title{Gets log-in information and sets up a connection to Oracle.}
\usage{
setupNrlsConn(db_user_name, db_conn_text, db_pass = NULL)
}
\arguments{
\item{db_user_name}{The user name to log into Oracle with}

\item{db_conn_text}{Database connection text}

\item{db_pass}{Password for user, if not provided with prompt user to provide}
}
\value{
An ROracle connection
}
\description{
Gets log-in information and sets up a connection to Oracle.
}
