% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyDb.r
\name{subsetLicences}
\alias{subsetLicences}
\title{Subset Licences}
\usage{
subsetLicences(licence_df, start_survey_date, end_survey_date)
}
\arguments{
\item{licence_df}{A data frame of all licences}

\item{start_survey_date}{start date of the survey (used to limit terms)}

\item{end_survey_date}{The date licence records are cut off at, this date
will be used to exclude licence records that start after this date.}
}
\value{
A data frame with the licence records limited to those aligning witha survey
}
\description{
Limit licences to those eligible for a survey
}
