% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyAnalysisLib.r
\name{summarizeLicPop}
\alias{summarizeLicPop}
\title{Summarize Licence Population}
\usage{
summarizeLicPop(elic_data, vendor_sales_data, survey_dates)
}
\arguments{
\item{elic_data}{Electronic licences data}

\item{vendor_sales_data}{Vendor sales of licences}

\item{survey_dates}{Survey date range}
}
\value{
A data frame summarized by licence categories and by electronic/vendor sales
}
\description{
Summarize the licence population totals using the E-Licence and Vendor Sales Data
}
