% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyAnalysisLib.r
\name{surveySpecificVariance}
\alias{surveySpecificVariance}
\title{Survey Specific Variance
Calculates the licence/draw type category specific variance.}
\usage{
surveySpecificVariance(
  survey_data,
  survey_response,
  lic_uniq_col_name = "licence_id",
  strata_col_names = c("draw_type", "resident_status", "licence_type", "age_category"),
  non_var_col_names = c("area", "method")
)
}
\arguments{
\item{survey_data}{All the survey result data for the month}

\item{survey_response}{A data frame of response totals by licence category/draw type}

\item{lic_uniq_col_name}{Column name identifying unique licences}

\item{strata_col_names}{Column names to stratify variances}

\item{non_var_col_names}{Columns that have no variance}
}
\value{
Variance values for the specific licence category/draw type in each
area and method
}
\description{
Survey Specific Variance
Calculates the licence/draw type category specific variance.
}
