% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyAnalysisLib.r
\name{surveyTotalVariance}
\alias{surveyTotalVariance}
\title{Survey Total Variance}
\usage{
surveyTotalVariance(
  specific_variance,
  sample_sizes,
  lic_total,
  lic_strata_col_names = c("draw_type", "resident_status", "licence_type",
    "age_category"),
  catch_strata_col_names = c("area", "method")
)
}
\arguments{
\item{specific_variance}{All the survey result data for the month}

\item{sample_sizes}{Sample sizes for each specific variance values}

\item{lic_total}{Total valid licences data frames}

\item{lic_strata_col_names}{Licence column names used to stratify}

\item{catch_strata_col_names}{Catch column names used to stratify}
}
\value{
Variance values for the specific licence category/draw type in each area and method
}
\description{
Combines the strata variances for the various licence categories/draw types to
to estimate the total catch variance for each area and method there is data for.
}
