% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyUtil.r
\name{validateValueDomain}
\alias{validateValueDomain}
\title{Validate Value domains}
\usage{
validateValueDomain(
  values,
  domain,
  error_message = "The following values are invalid:\\n\\n\%s\\n\\n"
)
}
\arguments{
\item{values}{The data that is to be validated}

\item{domain}{The set of valid values that "values" can take on.}

\item{error_message}{A template of a printf message for invalid values.}
}
\value{
The values provided
}
\description{
A helper function for validating data within a pre-defined set of possible values.
}
\note{
If the values vector contains values that are not in the domain vector, then the method calls stop
with an error message using the template provide in the "error_message" parameter.
}
