% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyAnalysisLib.r
\name{variance}
\alias{variance}
\title{This Variance function calculates variance base on a sample larger then provided for
in the data.  Values for all other observations are assumed to be zero and the variance calulation
handles all the additional zero values as a single term calculations.}
\usage{
variance(values, sample_size)
}
\arguments{
\item{values}{The values to calculate variance on, the length of this data vector is typically less then "n"}

\item{sample_size}{The date licence records are cut off at, this date will be used to exclude
licence records that start after this date.}
}
\value{
A data frame with the licence records loaded in a consistent R friendly format.
}
\description{
This Variance function calculates variance base on a sample larger then provided for
in the data.  Values for all other observations are assumed to be zero and the variance calulation
handles all the additional zero values as a single term calculations.
}
