% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyUtil.r
\name{writeProtectedCsv}
\alias{writeProtectedCsv}
\title{Write Protected CSV File}
\usage{
writeProtectedCsv(df, file_name)
}
\arguments{
\item{df}{Data frame to be written to CSV file}

\item{file_name}{File name to write CSV file to}
}
\description{
A helper function for writing CSV files and setting them to read-only for protection.
When writing the file, the file is first set to writable then written over then set to read-only
}
