% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyUtil.r
\name{getLicenceYearText}
\alias{getLicenceYearText}
\title{Get Licence Year Text}
\usage{
getLicenceYearText(lic_year)
}
\arguments{
\item{lic_year}{Licence year vector to convert to licence year text}
}
\value{
A vector of characters
}
\description{
Converts a Licence year (e.g 2016) into a licence year text (e.g. 2016-17).
Commonly used for displaying to users or the folder name of the data
}
