% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyUtil.r
\name{getTimeStampText}
\alias{getTimeStampText}
\title{Get Time Stamp Text}
\usage{
getTimeStampText()
}
\value{
A string with the current date and time
}
\description{
Provides a standardized text based time stamp for inclusion in file names
}
