% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyInstAnalysis.r
\name{runInstSurveyAnalysis}
\alias{runInstSurveyAnalysis}
\title{Run Instantaneous Selection Survey Analysis}
\usage{
runInstSurveyAnalysis(config, elic_data)
}
\arguments{
\item{config}{Configuration list}

\item{elic_data}{Electronic licence data frame}
}
\value{
list with survey analysis results
}
\description{
Analysis the responses to the iRec survey based on instantaneous selection method
integrated into the National Recreational Licence System (NRLS)
}
